added back memoization, refactored GrainReceiptTable
